python-api/requirements.txt
psycopg2-binary
Django
djangorestframework